+++ vocab/actor.test.ts
@@ -2,12 +2,15 @@ import {
   assert,
   assertEquals,
   assertFalse,
+  assertRejects,
   assertStrictEquals,
 } from "@std/assert";
 import * as fc from "fast-check";
+import * as mf from "mock_fetch";
 import {
   Actor,
   getActorClassByTypeName,
+  getActorHandle,
   getActorTypeName,
   isActor,
 } from "./actor.ts";
@@ -85,4 +88,57 @@ Deno.test("getActorClassByTypeName()", () => {
         ),
     ),
   );
+});
+
+Deno.test("getActorHandle()", async (t) => {
+  mf.install();
+
+  mf.mock(
+    "GET@/.well-known/webfinger",
+    (_) =>
+      new Response(
+        JSON.stringify({ subject: "acct:john@example.com" }),
+        { headers: { "Content-Type": "application/jrd+json" } },
+      ),
+  );
+
+  const actorId = new URL("https://foo.example.com/@john");
+  const actor = new Person({
+    id: actorId,
+    preferredUsername: "john",
+  });
+
+  await t.step("WebFinger subject", async () => {
+    assertEquals(await getActorHandle(actor), "@john@example.com");
+    assertEquals(await getActorHandle(actorId), "@john@example.com");
+  });
+
+  mf.mock(
+    "GET@/.well-known/webfinger",
+    (_) =>
+      new Response(
+        JSON.stringify({
+          subject: "https://foo.example.com/@john",
+          aliases: ["acct:john@bar.example.com"],
+        }),
+        { headers: { "Content-Type": "application/jrd+json" } },
+      ),
+  );
+
+  await t.step("WebFinger aliases", async () => {
+    assertEquals(await getActorHandle(actor), "@john@bar.example.com");
+    assertEquals(await getActorHandle(actorId), "@john@bar.example.com");
+  });
+
+  mf.mock(
+    "GET@/.well-known/webfinger",
+    (_) => new Response(null, { status: 404 }),
+  );
+
+  await t.step("no WebFinger", async () => {
+    assertEquals(await getActorHandle(actor), "@john@foo.example.com");
+    assertRejects(() => getActorHandle(actorId), TypeError);
+  });
+
+  mf.uninstall();
 });