+++ vocab/lookup.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals, assertInstanceOf } from "@std/assert";
-import * as mf from "https://deno.land/x/mock_fetch@0.3.0/mod.ts";
+import * as mf from "mock_fetch";
 import { mockDocumentLoader } from "../testing/docloader.ts";
 import { lookupObject } from "./lookup.ts";
 import { Object, Person } from "./vocab.ts";